+++ java/src/com/swiftnav/sbp/ndb/MsgNdbEvent.java
@@ -13,6 +13,8 @@
 
 package com.swiftnav.sbp.ndb;
 
+import java.math.BigInteger;
+
 import com.swiftnav.sbp.SBPMessage;
 import com.swiftnav.sbp.SBPBinaryException;
 import com.swiftnav.sbp.SBPStruct;
@@ -36,7 +38,7 @@ public class MsgNdbEvent extends SBPMessage {
 
     
     /** HW time in milliseconds. */
-    public long recv_time;
+    public BigInteger recv_time;
     
     /** Event type. */
     public int event;
